hivemind/requirements.txt
# HiveMind Phase 1 — Python dependencies
# No LLM libraries — Phase 1 is pure HMS connectivity and MCP scaffolding.

# MCP server framework
fastmcp>=2.0

# Environment variables
python-dotenv>=1.0.0

# Apache Thrift (Hive Metastore wire protocol)
thrift>=0.16.0

# Testing
pytest>=7.0
pytest-asyncio>=0.23
